Refactor tests (#342)
* Refactor tests

* Fix build tag

* Fix linting error
