+++ test/testkit/k8s/objects.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package k8s
 
 import (
