+++ test/testkit/k8s/service.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package k8s
 
 import (
