+++ test/testkit/k8s/verifiers/deployment.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package verifiers
 
 import (
