+++ test/testkit/kyma/istio/istio.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package istio
 
 var AccessLogAttributeKeys = []string{
