+++ test/testkit/kyma/telemetry/log/parser.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package log
 
 import (
