+++ test/testkit/kyma/telemetry/metric/pipeline.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package metric
 
 import (
