+++ test/testkit/kyma/telemetry/trace/pipeline.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package trace
 
 import (
