+++ test/testkit/matchers/metric_matchers.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package matchers
 
 import (
