+++ test/testkit/matchers/prometheus_matcher_test.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package matchers
 
 import (
