+++ test/testkit/matchers/trace_matchers_test.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package matchers
 
 import (
