+++ test/testkit/mocks/external_backend_service.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package mocks
 
 import (
