+++ test/testkit/mocks/log_spammer.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package mocks
 
 import (
