+++ test/testkit/otlp/metrics/exporter.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package metrics
 
 import (
