+++ test/testkit/otlp/traces/traces.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package traces
 
 import (
