+++ test/testkit/port_registry.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 package testkit
 
 type (
